+++ lib/pop3mail/date_converter.ex
@@ -19,9 +19,10 @@ defmodule Pop3mail.DateConverter do
      # Sometimes the day of the week is missing in the date. Fix that:
      # if date starts with digit, add the day first. We don't care which day it is, just add ???
      date_str =
-        case date_str =~ ~r/^\s?\d/ do
-           true  -> "???, " <> date_str
-           false -> date_str
+        if date_str =~ ~r/^\s?\d/ do
+           "???, " <> date_str
+        else
+           date_str
         end
      # httpd_util requires that single digit days have a leading zero. This is not always the case.
      day_and_date = date_str
@@ -29,9 +30,10 @@ defmodule Pop3mail.DateConverter do
                     |> String.trim_leading
      # add leading zero
      date_str =
-       case day_and_date =~ ~r/^\d\s/ do
-         true  -> String.slice(date_str, 0..4) <> "0" <> day_and_date
-         false -> date_str
+       if day_and_date =~ ~r/^\d\s/ do
+          String.slice(date_str, 0..4) <> "0" <> day_and_date
+       else
+          date_str
        end
      date_char_list = to_charlist(date_str)
      {{year, month, day}, {hour, minutes, seconds}} = :httpd_util.convert_request_date(date_char_list)